semi-usable build, selection and blank invoice adding
added scrolling buttons,
added some more checks and ranges to things in the code, so stuff does not error as much when put outside its limits, making more use of my clamp function i added.
added the ability to select invoices,
faster scrolling,
slightly changed my cache system,
added the ability to add blank invoices- editting them is to come in the future, this is good enough for now- for the sake of having a semi-usable build i can show to people 'n show my progress.
this was a nightmare to work on, and i didn't note much of my changes between pushes... but the progress shows, and i commented my code a lot - look to those.
also ignore addnewinvoice.php - it is work in progress and not used in the program rn, do not judge it.

also ignore the green border stuff, i was using it with testing to know what is where and the layering 'n such.

I also included some of my screenshots relating to when i was testing stuff, and in-between version changes